Factor out typechecking of toplevel
