+++ lang/elaborator/src/typechecker/mod.rs
@@ -1,4 +1,5 @@
 pub mod ctx;
+pub mod decls;
 pub mod subst;
 pub mod typecheck;
 pub mod util;
